# API-v2/swaggerUI/swaggerfile.json
{
	"swagger": "2.0",
	"info": {
		"title": "BetterSchool API",
		"description": "The api for BetterSchool",
		"version": "1.0.0"
	},
	"host": "api.betterschool.chph.tk",
	"basePath": "/",
	"schemes": ["https"],
	"paths": {
		"/schools": {
			"get": {
				"description": "Use this endpoint to get all the schools and their ID's",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Returns a list of schools and their respective ID's",
						"schema": {
							"type": "array",
							"items": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"example": "Amalieskram VGS"
									},
									"schoolID": {
										"type": "string",
										"example": "78f7ae51-b9e7-5c61-95e7-283b7cd7f8f2"
									}
								}
							},
							"xml": {
								"name": "main"
							}
						}
					}
				}
			}
		},
		"/school/{schoolID}/classes": {
			"get": {
				"description": "Use this endpoint to get a list of all classes in a school",
				"parameters": [
					{
						"name": "schoolID",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "The school-ID to get the classes from"
					}
				],
				"responses": {
					"200": {
						"description": "Returns a list of the classes registered to said school",
						"schema": {
							"type": "array",
							"items": {
								"type": "object",
								"properties": {
									"className": {
										"type": "string",
										"example": "3PBYA"
									},
									"classID": {
										"type": "string",
										"example": "f0907e85-8d20-51a2-a307-956b55ad77ad"
									}
								}
							},
							"xml": {
								"name": "main"
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/school/{schoolID}/class/{classID}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "schoolID",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "The ID of the school"
					},
					{
						"name": "classID",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "The ID of the class"
					}
				],
				"responses": {
					"200": {
						"description": "Returns a list of the classes registered to said school",
						"schema": {
							"type": "array",
							"items": {
								"type": "object",
								"properties": {
									"weekNr": {
										"type": "string",
										"example": "42"
									},
									"days": {
										"type": "array",
										"items": {
											"type": "object",
											"properties": {
												"name": {
													"type": "string",
													"example": "Fredag"
												},
												"date": {
													"type": "string",
													"example": "Fri Oct 21 2022"
												},
												"classes": {
													"type": "array",
													"items": {
														"type": "object",
														"properties": {
															"date": {
																"type": "string",
																"example": "21. oktober 2022"
															},
															"time": {
																"type": "string",
																"example": "13:15-14:15"
															},
															"room": {
																"type": "string",
																"example": "553"
															},
															"name": {
																"type": "string",
																"example": "Naturfag"
															},
															"teacher": {
																"type": "string",
																"example": "Ole-Kristian Eide"
															}
														}
													}
												}
											}
										}
									}
								}
							},
							"xml": {
								"name": "main"
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/addUser": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "obj",
						"in": "body",
						"description": "Credentials of the new user",
						"schema": {
							"type": "object",
							"properties": {
								"username": {
									"type": "string",
									"example": "e-bobkåre"
								},
								"pass": {
									"type": "string",
									"example": "BoblikerKål123"
								},
								"class": {
									"type": "string",
									"example": "2ITKA"
								},
								"schoolID": {
									"type": "string",
									"example": "78f7ae51-b9e7-5c61-95e7-283b7cd7f8f2"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Retuns 400 if the body object did not include the correct params"
					},
					"401": {
						"description": "Returns 401 if the api failed to verify the credentials"
					}
				}
			}
		}
	}
}